cli: move argument parser to its own file
